agents/rahulmx_agent/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
pydantic==2.12.5
google-adk==1.29.0
google-genai==1.72.0
python-dotenv==1.2.2
